requirements.txt
pydds
Pillow
pyinstaller
requests